migration of changes into ZHighLayoutPropagation
