+++ src/Accelerators/NNPA/Transform/ZHigh/ZHighLayoutPropagation.cpp
@@ -4,7 +4,7 @@
 
 //===---------- ZHighLayoutPropagation.cpp - ZHigh High Level Optimizer ---===//
 //
-// Copyright 2019-2022 The IBM Research Authors.
+// Copyright 2019-2025 The IBM Research Authors.
 //
 // =============================================================================
 //
@@ -24,10 +24,33 @@
 #include "src/Accelerators/NNPA/Pass/NNPAPasses.hpp"
 #include "src/Accelerators/NNPA/Support/LayoutHelper.hpp"
 #include "src/Dialect/ONNX/ONNXOps.hpp"
+#include "src/Support/TypeUtilities.hpp"
 
 using namespace mlir;
 using namespace onnx_mlir;
 using namespace onnx_mlir::zhigh;
+
+#define DEBUG_TYPE "zhigh-layout-propagation"
+
+// If set to 1, enable multiple distinct layouts to elementwise compute
+// operations; 0 otherwise. We can support the "compiler supported" layouts
+// because we only care about SIMD gen of the E1 (innermost) dimension.
+#define ELEMENTWISE_WITH_MULTIPLE_LAYOUTS 1
+
+static bool canOpFuseWithStickUnstick(Operation *op) {
+  if (!op)
+    return false;
+  // Exceptions:
+  // o no cast as they may have different input sizes/output sizes.
+  if (isa<ONNXCastOp>(op))
+    return false;
+    // Return true for all of the elementwise operations.
+#define ELEMENTWISE_ALL(_OP_TYPE)                                              \
+  if (isa<_OP_TYPE>(op))                                                       \
+    return true;
+#include "src/Conversion/ONNXToKrnl/Math/Elementwise.hpp"
+  return false;
+}
 
 namespace onnx_mlir {
 namespace zhigh {
@@ -93,128 +116,6 @@ Type getZTensorType(
 //===----------------------------------------------------------------------===//
 // ZHigh layout propagation patterns
 //===----------------------------------------------------------------------===//
-
-//===----------------------------------------------------------------------===//
-// Layout propagation for ONNX unary element-wise operations
-//
-// ONNX unary operations, such as Relu and Tanh, do not depend on a specific
-// layout. Thus, it is possible to use the layout of the previous zTensor. In
-// other words, propagate the existing layout down to the unary operation, so
-// that other rules can be applied to remove unstick-stick pairs.For example,
-// This sequence of operations:
-//   zhigh.Conv -> Unstick (NHWC) -> onnx.Relu
-// will become:
-//   zhigh.Conv -> onnx.Relu -> Unstick (NHWC)
-// then, canonicalization of unstick/stick will remove the unstick-stick pair.
-//
-//===----------------------------------------------------------------------===//
-
-template <typename ONNX_OP>
-class ONNXUnaryOpLayoutPropPattern : public OpRewritePattern<ONNX_OP> {
-public:
-  using OpRewritePattern<ONNX_OP>::OpRewritePattern;
-
-  LogicalResult matchAndRewrite(
-      ONNX_OP unaryOp, PatternRewriter &rewriter) const override {
-    Operation *genericOp = unaryOp.getOperation();
-    Location loc = genericOp->getLoc();
-
-    Value input = unaryOp.getX();
-    Value output = unaryOp.getY();
-
-    // Input is a block argument, do nothing.
-    if (mlir::dyn_cast<BlockArgument>(input))
-      return failure();
-
-    // Input is a CPU tensor, do nothing.
-    auto unstickOp = mlir::dyn_cast<ZHighUnstickOp>(input.getDefiningOp());
-    if (!unstickOp)
-      return failure();
-
-    // Input is unstickified from a zTensor. Do computation directly on the
-    // zTensor.
-    Value zTensor = unstickOp.getIn();
-    Value zOutput = rewriter.create<ONNX_OP>(loc, zTensor.getType(), zTensor);
-    Value replacedValue =
-        rewriter.create<ZHighUnstickOp>(loc, output.getType(), zOutput);
-    rewriter.replaceOp(genericOp, replacedValue);
-    return success();
-  }
-};
-
-//===----------------------------------------------------------------------===//
-// Layout propagation for ONNX binary element-wise operations
-//
-// ONNX binary operations, such as Add and Div, do not depend on a specific
-// layout. Thus, it is possible to use the layout of the previous zTensor. In
-// other words, propagate the existing layout down to the unary operation, so
-// that other rules can be applied to remove unstick-stick pairs.For example,
-// This sequence of operations:
-//   X -> Unstick (NHWC) -> onnx.Add
-//                              ^
-//   Y -> Unstick (NHWC) -------|
-//
-// will become:
-//   X -> onnx.Add -> Unstick (NHWC)
-//          ^
-//   Y -----|
-//
-// then, canonicalization of unstick/stick will remove the unstick-stick pairs.
-//
-//===----------------------------------------------------------------------===//
-
-template <typename ONNX_OP>
-class ONNXBinaryOpLayoutPropPattern : public OpRewritePattern<ONNX_OP> {
-public:
-  using OpRewritePattern<ONNX_OP>::OpRewritePattern;
-
-  LogicalResult matchAndRewrite(
-      ONNX_OP binaryOp, PatternRewriter &rewriter) const override {
-    Operation *genericOp = binaryOp.getOperation();
-    Location loc = genericOp->getLoc();
-
-    Value A = binaryOp.getA();
-    Value B = binaryOp.getB();
-    Value output = binaryOp.getC();
-
-    // Input is a block argument, do nothing.
-    if (mlir::dyn_cast<BlockArgument>(A) || mlir::dyn_cast<BlockArgument>(B))
-      return failure();
-
-    // Input is a CPU tensor, do nothing.
-    auto unstickAOp = mlir::dyn_cast<ZHighUnstickOp>(A.getDefiningOp());
-    auto unstickBOp = mlir::dyn_cast<ZHighUnstickOp>(B.getDefiningOp());
-    if (!unstickAOp || !unstickBOp)
-      return failure();
-
-    // Input is unstickified from a zTensor. Do computation directly on the
-    // zTensor.
-    Value zTensorA = unstickAOp.getIn();
-    Value zTensorB = unstickBOp.getIn();
-    Type zTensorAType = zTensorA.getType();
-    Type zTensorBType = zTensorB.getType();
-    ZTensorEncodingAttr::DataLayout ALayout = getZTensorLayout(zTensorAType);
-    ZTensorEncodingAttr::DataLayout BLayout = getZTensorLayout(zTensorBType);
-
-    // Support same layout for zTensor A & B in the current implementation.
-    // TODO: Support arbitrary layout in the future. In such a case, what should
-    // be the output layout?
-    if ((ALayout != ZTensorEncodingAttr::DataLayout::UNDEFINED) &&
-        (ALayout != BLayout))
-      return failure();
-
-    // Construct the zTensor type from the output CPU tensor.
-    Type zOutputType = getZTensorType(rewriter, loc, output,
-        convertZTensorDataLayoutToStringAttr(rewriter, ALayout));
-
-    Value zOutput =
-        rewriter.create<ONNX_OP>(loc, zOutputType, zTensorA, zTensorB);
-    Value replacedValue =
-        rewriter.create<ZHighUnstickOp>(loc, output.getType(), zOutput);
-    rewriter.replaceOp(genericOp, replacedValue);
-    return success();
-  }
-};
 
 /// The pattern
 ///   onnx.Concat (zhigh.Unstick (%X1), zhigh.Unstick (%X2)) { axis })
@@ -312,6 +213,311 @@ private:
   }
 };
 
+//===----------------------------------------------------------------------===//
+// ZHigh layout propagation patterns for elementwise ops
+//===----------------------------------------------------------------------===//
+
+using OperationSet = std::set<Operation *>;
+
+#if !ELEMENTWISE_WITH_MULTIPLE_LAYOUTS
+// Make sure that all inputs have either an undefined layout or the same as
+// reference layout.
+static bool suitableLayout(
+    Operation *op, ZTensorEncodingAttr::DataLayout refLayout) {
+  // Now iterate over each of the inputs to op.
+  for (Value v : op->getOperands()) {
+    // Check if we have a layout and if it is compatible.
+    ZTensorEncodingAttr::DataLayout vLayout =
+        onnx_mlir::zhigh::getZTensorLayout(v.getType());
+    if (vLayout == ZTensorEncodingAttr::DataLayout::UNDEFINED ||
+        vLayout == refLayout)
+      continue;
+    // We have a Z layout and its not the same, abort
+    return false;
+  }
+  return true;
+}
+#endif
+
+// Make sure that all inputs and outputs have the right element type. Currently
+// only support f32 or d.
+static bool suitableComputeType(Type type) {
+  Type elementType = getElementTypeOrSelf(type);
+  if (elementType.isF32())
+    return true;
+  if (elementType.isF16() && isZTensor(type))
+    return true;
+  return false;
+}
+
+static bool suitableComputeType(Operation *op) {
+  for (Value v : op->getOperands()) {
+    if (!suitableComputeType(v.getType()))
+      return false;
+  }
+  for (Value v : op->getResults()) {
+    if (!suitableComputeType(v.getType()))
+      return false;
+  }
+  return true;
+}
+
+// Give an op that consumes referenceInputVal, check if all the other input
+// values have either the same shape for their last dim, or if there is a
+// broadcast known at compile time.
+bool sameLastDimOrStaticBroadcast(
+    DimAnalysis *dimAnalysis, Operation *op, Value referenceInputVal) {
+  // Get innermost shape of reference input val.
+  ShapedType refType = mlir::dyn_cast<ShapedType>(referenceInputVal.getType());
+  if (!refType)
+    return false; // Expected shaped type, abort.
+  if (refType.getRank() == 1)
+    return true; // Unary ops never have broadcasts.
+  int64_t innermostShapeOfRef = getShape(refType, -1);
+  // Now iterate over each of the inputs to op.
+  for (Value v : op->getOperands()) {
+    // Same dimension, we are fine.
+    if (v == referenceInputVal ||
+        dimAnalysis->sameDim(v, -1, referenceInputVal, -1))
+      continue;
+    // Check if we have a uni-directional broadcast known at compile time.
+    ShapedType vType = mlir::dyn_cast<ShapedType>(v.getType());
+    if (!vType)
+      return false;
+    int64_t innermostShapeOfV = getShape(vType, -1);
+    if (!ShapedType::isDynamic(innermostShapeOfRef) && innermostShapeOfV == 1)
+      continue;
+    if (!ShapedType::isDynamic(innermostShapeOfV) && innermostShapeOfRef == 1)
+      continue;
+    // Not the same size and no uni-broadcasting at compile time, fail.
+    return false;
+  }
+  return true;
+}
+
+// When value is consumed by only one op, returns that op. Otherwise return
+// null.
+Operation *getSingleUseOperationOf(Value val) {
+  Operation *singleOp = nullptr;
+  int multipleComputeOpNum = 0;
+  for (OpOperand &use : val.getUses()) {
+    Operation *useOp = use.getOwner();
+    if (singleOp == nullptr)
+      singleOp = useOp;
+    else if (singleOp != useOp) {
+      multipleComputeOpNum++;
+      LLVM_DEBUG({
+        if (multipleComputeOpNum == 1) {
+          // TODO, could look into tolerating multiple "supported" ops + some
+          // free ops such as "DimOp". Printout here highlights the missing
+          // opportunities.
+          llvm::dbgs() << "Unstick -> multiple compute ops fusion FAILURE:\n  ";
+          val.dump();
+          llvm::dbgs() << "  ";
+          singleOp->dump();
+        }
+        llvm::dbgs() << "  ";
+        useOp->dump();
+      });
+    }
+  }
+  return multipleComputeOpNum ? nullptr : singleOp;
+}
+
+static void explanation(
+    Operation *computeOp, ZHighUnstickOp unstickOp, std::string message) {
+  llvm::dbgs() << "Unstick->compute (" << computeOp->getName() << ") fusion "
+               << message << ":\n  ";
+  unstickOp.dump();
+  llvm::dbgs() << "  ";
+  computeOp->dump();
+}
+
+static void explanation(
+    Operation *computeOp, ZHighStickOp stickOp, std::string message) {
+  llvm::dbgs() << "Compute->stick (" << computeOp->getName() << ") fusion "
+               << message << ":\n  ";
+  computeOp->dump();
+  llvm::dbgs() << "  ";
+  stickOp.dump();
+}
+
+// Return compute operation when fusion of unstick -> compute can be fused,
+// Nullptr is returned when not feasible. If unstick -> compute -> stick is
+// detected, then stickOp is further defined, otherwise it is  nullptr,
+Operation *patternForFusionFromUnstick(
+    ZHighUnstickOp unstickOp, DimAnalysis *dimAnalysis) {
+  Value unstickInVal = unstickOp.getIn();
+  Value unstickOutVal = unstickOp.getOut();
+  // For merge, unstick value can only be used only once.
+  Operation *computeOp = getSingleUseOperationOf(unstickOutVal);
+  // Supported compute op?
+  if (!computeOp)
+    return nullptr;
+  if (!canOpFuseWithStickUnstick(computeOp)) {
+    LLVM_DEBUG(/* usefull to find new opportunities not supported yet*/
+        explanation(computeOp, unstickOp, "FAILURE compute op cannot fuse"));
+    return nullptr;
+  }
+  if (!suitableComputeType(computeOp)) {
+    LLVM_DEBUG(explanation(
+        computeOp, unstickOp, "FAILURE due to non f32/dlf16 element type"));
+    return nullptr;
+  }
+  // We must support this layout.
+  if (isZTensor(unstickInVal.getType()) &&
+      !supportedLayoutForCompilerGeneratedStickUnstick(unstickInVal)) {
+    LLVM_DEBUG(
+        explanation(computeOp, unstickOp, "FAILURE due to unstick shape"));
+    return nullptr;
+  }
+  // Suitable shapes?
+  if (!sameLastDimOrStaticBroadcast(dimAnalysis, computeOp, unstickOutVal)) {
+    LLVM_DEBUG(
+        explanation(computeOp, unstickOp, "FAILURE due to input shapes"));
+    return nullptr;
+  }
+#if !ELEMENTWISE_WITH_MULTIPLE_LAYOUTS
+  // Suitable layout?
+  ZTensorEncodingAttr::DataLayout unstickLayout =
+      onnx_mlir::zhigh::getZTensorLayout(unstickInVal.getType());
+  if (!suitableLayout(computeOp, unstickLayout)) {
+    LLVM_DEBUG(explanation(
+        computeOp, unstickOp, "FAILURE due to input zTensor layouts"));
+    return nullptr;
+  }
+#endif
+  // Success.
+  LLVM_DEBUG(explanation(computeOp, unstickOp, ""));
+  return computeOp;
+}
+
+// Return compute operation when compute -> stick can be fused.
+Operation *patternForFusionFromStick(
+    ZHighStickOp stickOp, DimAnalysis *dimAnalysis) {
+  Value stickInVal = stickOp.getIn();
+  Value stickOutVal = stickOp.getOut();
+  // Input of stick can only be used once.
+  if (!stickInVal.hasOneUse())
+    return nullptr;
+  // Get use operation and ensure it can be used.
+  Operation *computeOp = stickInVal.getDefiningOp();
+  if (!computeOp)
+    return nullptr;
+  if (!canOpFuseWithStickUnstick(computeOp)) {
+    LLVM_DEBUG(/* usefull to find new opportunities not supported yet*/
+        explanation(computeOp, stickOp, "FAILURE compute op cannot fuse"));
+    return nullptr;
+  }
+  if (!suitableComputeType(computeOp)) {
+    LLVM_DEBUG(explanation(
+        computeOp, stickOp, "FAILURE due to non f32/dlf16 element type"));
+    return nullptr;
+  }
+  // We must support this layout.
+  if (isZTensor(stickOutVal.getType()) &&
+      !supportedLayoutForCompilerGeneratedStickUnstick(stickOutVal)) {
+    LLVM_DEBUG(explanation(computeOp, stickOp, "FAILURE due to stick layout"));
+    return nullptr;
+  }
+#if !ELEMENTWISE_WITH_MULTIPLE_LAYOUTS
+  ZTensorEncodingAttr::DataLayout stickLayout =
+      onnx_mlir::zhigh::getZTensorLayout(stickOp.getOut().getType());
+  if (!suitableLayout(computeOp, stickLayout)) {
+    LLVM_DEBUG(explanation(
+        computeOp, stickOp, "FAILURE due to input zTensor layouts"));
+    return nullptr;
+  }
+#endif
+  // ZHighUnstickOp unstickOp = computeOp
+  LLVM_DEBUG(explanation(computeOp, stickOp, ""));
+  return computeOp;
+}
+
+class PatternsStartingFromUnstick : public OpRewritePattern<ZHighUnstickOp> {
+public:
+  DimAnalysis *dimAnalysis;
+  OperationSet *processedUnstick;
+
+  PatternsStartingFromUnstick(MLIRContext *context, DimAnalysis *dimAnalysis,
+      OperationSet *processedUnstick)
+      : OpRewritePattern<ZHighUnstickOp>(context, 1), dimAnalysis(dimAnalysis),
+        processedUnstick(processedUnstick) {}
+
+  using OpRewritePattern<ZHighUnstickOp>::OpRewritePattern;
+  LogicalResult matchAndRewrite(
+      ZHighUnstickOp unstickOp, PatternRewriter &rewriter) const override {
+
+    // Process each stick/unstick only once.
+    if (processedUnstick->find(unstickOp.getOperation()) !=
+        processedUnstick->end())
+      return failure();
+
+    Operation *computeOp = patternForFusionFromUnstick(unstickOp, dimAnalysis);
+    if (computeOp) {
+      int64_t operandNum = computeOp->getNumOperands();
+      rewriter.modifyOpInPlace(computeOp, [&]() {
+        for (int64_t i = 0; i < operandNum; ++i)
+          if (computeOp->getOperand(i) == unstickOp.getOut())
+            // Have to replace this operand by input of unstick.
+            computeOp->setOperand(i, unstickOp.getIn());
+      });
+      processedUnstick->insert(unstickOp.getOperation());
+      return success();
+    }
+    return failure();
+  }
+};
+
+class PatternsEndingWithStick : public OpRewritePattern<ZHighStickOp> {
+public:
+  DimAnalysis *dimAnalysis;
+  OperationSet *processedStick;
+
+  PatternsEndingWithStick(MLIRContext *context, DimAnalysis *dimAnalysis,
+      OperationSet *processedStick)
+      : OpRewritePattern<ZHighStickOp>(context, 1), dimAnalysis(dimAnalysis),
+        processedStick(processedStick) {}
+
+  using OpRewritePattern<ZHighStickOp>::OpRewritePattern;
+  LogicalResult matchAndRewrite(
+      ZHighStickOp stickOp, PatternRewriter &rewriter) const override {
+
+    // Process each stick/unstick only once.
+    if (processedStick->find(stickOp.getOperation()) != processedStick->end())
+      return failure();
+
+    Operation *computeOp = patternForFusionFromStick(stickOp, dimAnalysis);
+    if (computeOp) {
+      int64_t resultNum = computeOp->getNumResults();
+      assert(resultNum == 1 && "expect only one result for any fused ops");
+      assert(computeOp->getResult(0) == stickOp.getIn() &&
+             "expected ouput to be stick input");
+      // New compute op: has type of the stick output.
+      auto newResultType =
+          llvm::dyn_cast<RankedTensorType>(stickOp.getOut().getType());
+      // Clone compute state, insert in regions, and create.
+      OperationState state(computeOp->getLoc(),
+          computeOp->getName().getStringRef(), computeOp->getOperands(),
+          {newResultType}, computeOp->getAttrs());
+      for (unsigned i = 0, e = computeOp->getNumRegions(); i < e; ++i)
+        state.addRegion();
+      Operation *newComputeOp = rewriter.create(state);
+      // Where ever stick outputs were used, now it should be the new
+      // compute's result.
+      rewriter.replaceOp(stickOp.getOperation(), newComputeOp->getResult(0));
+      // Keep track of the removed ops.
+      processedStick->insert(stickOp.getOperation());
+      return success();
+    }
+    return failure();
+  }
+};
+
+//===----------------------------------------------------------------------===//
+// ZHigh layout propagation from table gen.
+//===----------------------------------------------------------------------===//
+
 /// Use anonymous namespace to avoid duplication symbol `populateWithGenerated`
 /// among multiple tablegen-based definitions.
 
@@ -323,8 +529,7 @@ private:
 //===----------------------------------------------------------------------===//
 
 struct ZHighLayoutPropagationPass
-    : public PassWrapper<ZHighLayoutPropagationPass,
-          OperationPass<func::FuncOp>> {
+    : public PassWrapper<ZHighLayoutPropagationPass, OperationPass<ModuleOp>> {
 
   MLIR_DEFINE_EXPLICIT_INTERNAL_INLINE_TYPE_ID(ZHighLayoutPropagationPass)
 
@@ -334,46 +539,56 @@ struct ZHighLayoutPropagationPass
     return "Layout propagation at ZHighIR.";
   }
 
+  ZHighLayoutPropagationPass() = default;
+  ZHighLayoutPropagationPass(const ZHighLayoutPropagationPass &pass)
+      : PassWrapper<ZHighLayoutPropagationPass, OperationPass<ModuleOp>>() {}
+  ZHighLayoutPropagationPass(bool disableElementwiseLayoutProp) {
+    this->disableElementwiseLayoutProp = disableElementwiseLayoutProp;
+  }
+
+  Option<bool> disableElementwiseLayoutProp{*this,
+      "disable-elementwise-layout-propagation",
+      llvm::cl::desc("Disable the propagation of ZHigh layouts into "
+                     "elementwise operations"),
+      llvm::cl::init(false)};
+
   void runOnOperation() override {
-    Operation *function = getOperation();
+    ModuleOp module = getOperation();
     ConversionTarget target(getContext());
     RewritePatternSet patterns(&getContext());
 
     // Layout propagation for ZHigh Ops.
     populateWithGenerated(patterns);
 
-    // Layout propagation for ONNX Ops. Since DLFLOAT16 is not supported in
-    // LLVM, we disable computation operators like Add, Div, etc. and enable
-    // only data movement operators like Concat.
-
-    // Add
-    // patterns.insert<ONNXBinaryOpLayoutPropPattern<ONNXAddOp>>(&getContext());
-    // Div
-    // patterns.insert<ONNXBinaryOpLayoutPropPattern<ONNXDivOp>>(&getContext());
-    // Mul
-    // patterns.insert<ONNXBinaryOpLayoutPropPattern<ONNXMulOp>>(&getContext());
-    // Sub
-    // patterns.insert<ONNXBinaryOpLayoutPropPattern<ONNXSubOp>>(&getContext());
-    // Reciprocal
-    // patterns.insert<ONNXUnaryOpLayoutPropPattern<ONNXReciprocalOp>>(
-    //     &getContext());
-    // Sqrt
-    // patterns.insert<ONNXUnaryOpLayoutPropPattern<ONNXSqrtOp>>(&getContext());
-
     // Concat
     patterns.insert<ONNXConcatLayoutPropagatePattern>(&getContext());
+
+    // Elementwise operations driven from stick/unstick.
+    DimAnalysis *dimAnalysis = nullptr;
+    OperationSet processedStickUnstick;
+
+    if (!disableElementwiseLayoutProp) {
+      dimAnalysis = new DimAnalysis(module);
+      dimAnalysis->analyze();
+      patterns.insert<PatternsStartingFromUnstick>(
+          &getContext(), dimAnalysis, &processedStickUnstick);
+      patterns.insert<PatternsEndingWithStick>(
+          &getContext(), dimAnalysis, &processedStickUnstick);
+    }
 
     // We want to canonicalize stick/unstick ops during this pass to simplify
     // rules in this pass.
     ZHighStickOp::getCanonicalizationPatterns(patterns, &getContext());
     ZHighUnstickOp::getCanonicalizationPatterns(patterns, &getContext());
-    (void)applyPatternsGreedily(function, std::move(patterns));
+    (void)applyPatternsGreedily(module, std::move(patterns));
   }
 };
 } // anonymous namespace
 
-std::unique_ptr<Pass> createZHighLayoutPropagationPass() {
-  return std::make_unique<ZHighLayoutPropagationPass>();
+std::unique_ptr<Pass> createZHighLayoutPropagationPass(
+    bool disableElementwiseLayoutProp) {
+  return std::make_unique<ZHighLayoutPropagationPass>(
+      disableElementwiseLayoutProp);
 }
 
 } // namespace zhigh
